Add tests for err-serializer
